+++ src/lua/shared/openutils/logutils.lua
@@ -5,6 +5,7 @@
 
 logutils = {
     version = "0.13.0", -- in semantic versioning (http://semver.org/)
+    pzversion = getCore():getVersionNumber(),
 
     -- Placeholders for Project Zomboid log file names.
     -- Project Zomboid generates files like this 24-08-19_18-11_chat.txt