+++ CHANGELOG.md
@@ -8,6 +8,9 @@ All notable changes to this project will be documented in this file.
 - Added skipping player tick log when character is dead.
 - Added death log to player file.
 
+## Removed
+- Removed functions writeLog, getLogLinePrefix, getLocation, getPlayerSafehouses, getPlayerPerks, getPlayerTraits, getPlayerStats, getPlayerHealth, getVehicleInfo from LogExtenderClient. These functions were deprecated on v0.12 and moved to logutils.
+
 ## [v0.12.0] - 2024-06-03
 ### Fixed
 - Fixed empty objName value in map_alternative log record. When player picked up IsoThumpable, placed and destroyed it LogExtender added line with empty objName value.